Refactor BuildChange into bsp package
